feat: improve contract structure
