+++ contracts/scheme.tlb
@@ -42,7 +42,8 @@ burn query_id:uint64 amount:(VarUInteger 16)
        response_destination:MsgAddress custom_payload:(Maybe ^Cell)
        = InternalMsgBody;
 
-storage#_ jetton_balances: (HashmapE 256 (VarUInteger 16)) receiver_address:MsgAddressInt unlocked_at:uint64 owner_address:MsgAddressInt = Storage;
+storage#_ jetton_balances: (HasmapE 256 Coins) owner_address:MsgIntAddress receiver_address:MsgIntAddress unlocked_at:uint64 = Storage;
+init query_id:uint64 supported_jettons: (HashmapE 256 Coins) = InternalMsgBody;
 lockup_data query_id:uint64 = InternalMsgBody;
 edit_owner query_id:uint64 owner:MsgAddress = InternalMsgBody;
 edit_receiver query_id:uint64 owner:MsgAddress = InternalMsgBody;
