+++ wrappers/JettonLockup.ts
@@ -1,4 +1,4 @@
-import { Address, beginCell, Cell, Contract, contractAddress, ContractProvider, Sender, SendMode, toNano } from 'ton-core';
+import { Address, beginCell, Cell, Contract, contractAddress, ContractProvider, Sender, SendMode, toNano, Dictionary, DictionaryValue } from 'ton-core';
 
 export type JettonLockupConfig = {
     owner: Address;
@@ -6,12 +6,28 @@ export type JettonLockupConfig = {
     unlockedAt: number;
 };
 
+export type BalanceValue = {
+    balance: bigint
+};
+
+export const BalancesValue: DictionaryValue<BalanceValue> = {
+    serialize: (src: BalanceValue, builder) => {
+        builder
+            .storeCoins(src.balance)
+    },
+    parse: (src) => {
+        return {
+            balance: src.loadCoins(),
+        }
+    },
+};
+
 export function jettonLockupConfigToCell(config: JettonLockupConfig): Cell {
     return beginCell()
         .storeUint(0, 1)
+        .storeAddress(config.owner)
         .storeAddress(config.receiver)
         .storeUint(Math.floor(config.unlockedAt), 64)
-        .storeAddress(config.owner)
         .endCell();
 }
 
@@ -28,17 +44,21 @@ export class JettonLockup implements Contract {
         return new JettonLockup(contractAddress(workchain, init), init);
     }
 
-    async sendDeploy(provider: ContractProvider, via: Sender) {
+    async sendDeploy(provider: ContractProvider, via: Sender, supportedWallets: Address[]) {
+        let jettonBalances = Dictionary.empty(Dictionary.Keys.Buffer(32), BalancesValue); 
+        for (let _wallet of supportedWallets)
+            jettonBalances.set(_wallet.hash, { balance: toNano('0') });
+
         await provider.internal(via, {
-            value: toNano('0.5'),
+            value: toNano('0.2'),
             sendMode: SendMode.PAY_GAS_SEPARATLY,
-            body: beginCell().storeUint(0xd53276db, 32).storeUint(0, 64).endCell(),
+            body: beginCell().storeUint(0, 32 + 64).storeDict(jettonBalances).endCell(),
         });
     }
 
-    async getLockupData(provider: ContractProvider): Promise<[Address, Address, number, Cell]> {
+    async getLockupData(provider: ContractProvider): Promise<[Address, Address, number]> {
         const { stack } = await provider.get("lockup_data", []);
-        return [stack.readAddress(), stack.readAddress(), stack.readNumber(), stack.readCell()];
+        return [stack.readAddress(), stack.readAddress(), stack.readNumber()];
     }
 
     async sendTransfer(provider: ContractProvider, via: Sender, to: Address) {
